Move capabilities onto separate request unions.
[ci skip-rust]
[ci skip-build-wheels]
