+++ src/python/pants/bsp/util_rules/compile.py
@@ -7,44 +7,25 @@ import time
 import uuid
 from collections import defaultdict
 from dataclasses import dataclass
-from typing import ClassVar, Generic, Type, TypeVar
+from typing import Type, TypeVar
 
 from pants.bsp.context import BSPContext
 from pants.bsp.protocol import BSPHandlerMapping
 from pants.bsp.spec.base import BuildTargetIdentifier, StatusCode, TaskId
 from pants.bsp.spec.compile import CompileParams, CompileReport, CompileResult, CompileTask
 from pants.bsp.spec.task import TaskFinishParams, TaskStartParams
-from pants.bsp.util_rules.targets import BSPBuildTargetInternal
+from pants.bsp.util_rules.targets import BSPBuildTargetInternal, BSPCompileRequest, BSPCompileResult
 from pants.engine.fs import Workspace
 from pants.engine.internals.native_engine import EMPTY_DIGEST, Digest, MergeDigests
 from pants.engine.internals.selectors import Get, MultiGet
 from pants.engine.rules import _uncacheable_rule, collect_rules
 from pants.engine.target import FieldSet, Targets
-from pants.engine.unions import UnionMembership, UnionRule, union
+from pants.engine.unions import UnionMembership, UnionRule
 from pants.util.ordered_set import FrozenOrderedSet
 
 _logger = logging.getLogger(__name__)
 
 _FS = TypeVar("_FS", bound=FieldSet)
-
-
-@union
-@dataclass(frozen=True)
-class BSPCompileRequest(Generic[_FS]):
-    """Hook to allow language backends to compile targets."""
-
-    field_set_type: ClassVar[Type[_FS]]
-
-    bsp_target: BSPBuildTargetInternal
-    field_sets: tuple[_FS, ...]
-
-
-@dataclass(frozen=True)
-class BSPCompileResult:
-    """Result of compilation of a target capable of target compilation."""
-
-    status: StatusCode
-    output_digest: Digest
 
 
 class CompileRequestHandlerMapping(BSPHandlerMapping):
